+++ plugins/TabSRMM/src/sidebar.cpp
@@ -148,7 +148,7 @@ const SIZE& CSideBarButton::measureItem()
 		HFONT oldFont = reinterpret_cast<HFONT>(::SelectObject(dc, ::GetStockObject(DEFAULT_GUI_FONT)));
 
 		mir_sntprintf(tszLabel, 255, _T("%s"), m_dat->newtitle);
-		::GetTextExtentPoint32(dc, tszLabel, mir_tstrlen(tszLabel), &sz);
+		::GetTextExtentPoint32(dc, tszLabel, (int)mir_tstrlen(tszLabel), &sz);
 
 		sz.cx += 28;
 		if (m_dat->pContainer->dwFlagsEx & TCF_CLOSEBUTTON)
@@ -1180,9 +1180,9 @@ const SIZE& __fastcall CSideBar::m_measureAdvancedVertical(CSideBarButton* item)
 		HDC	dc = ::GetDC(dat->hwnd);
 
 		HFONT	hOldFont = reinterpret_cast<HFONT>(::SelectObject(dc, CInfoPanel::m_ipConfig.hFonts[IPFONTID_NICK]));
-		::GetTextExtentPoint32(dc, dat->cache->getNick(), mir_tstrlen(dat->cache->getNick()), &szFirstLine);
+		::GetTextExtentPoint32(dc, dat->cache->getNick(), (int)mir_tstrlen(dat->cache->getNick()), &szFirstLine);
 		::SelectObject(dc, CInfoPanel::m_ipConfig.hFonts[IPFONTID_STATUS]);
-		::GetTextExtentPoint32(dc, dat->szStatus, mir_tstrlen(dat->szStatus), &szSecondLine);
+		::GetTextExtentPoint32(dc, dat->szStatus, (int)mir_tstrlen(dat->szStatus), &szSecondLine);
 		::SelectObject(dc, hOldFont);
 		ReleaseDC(dat->hwnd, dc);
